Remove assisted inject
It has been upstreamed into Dagger. This project will now only host inflation inject.
