--- .travis.yml
@@ -1,26 +1,0 @@
-language: android
-
-jdk:
-  - openjdk8
-
-before_install:
-  - mkdir "$ANDROID_HOME/licenses" || true
-  - echo "d56f5187479451eabf01fb78af6dfcb131a6481e" > "$ANDROID_HOME/licenses/android-sdk-license"
-  - echo "24333f8a63b6825ea9c5514f83c2829b004d1fee" >> "$ANDROID_HOME/licenses/android-sdk-license"
-
-after_success:
-  - .buildscript/deploy_snapshot.sh
-
-env:
-  global:
-    - secure: "JhAOJgP2RFuJQm7wRtyJcfMkQuDXMk2K7jxLtpYWIIbH/NAtw9/0S8299YqeTa2nDlvIa40kPrJGiCkuXlUqop+lQKvl0s6K43JmvuzNMcROFSU/Ev6+Kn7Smj9AA0Iscrjodapc7gSXEILGbQt+Ld8raAOS+MmORDh37QPOSREWZD+4EKRdlynnAciqnp2ZK7lJu8TTyU0mXf0zKnuk1m2mkeLe5dMCTV+L5yxT2vmG91S5EGQthg+NNu1V1cW8MxhV58duBrW0CHt/5cfEDWCTi1IzOd1EsRGBBLk9WiJ3Lx5lIaZ9CypCPYKlbz8ED0IrR1SMQIxSteByCNMAyPdu37gsS5HWnaphTYX3E23mIOCvbh/42YI0vOOQViuPQpsHd3ygo7ZS7sDDhO6gbcRsSZMMbVs4yYS/8eZLHBk7o6h9cURG039WDJ2kGGAGkFQ3hQmZLdr5FWXVb+VjaQ7DgH+Io4xdg73hpSOXpO0EFOHlVTyz8yfdVQlQuXdd+zwe7GOuQozlNkGgDyLiKODaldHXN2AlKmWaAoRVQELmQiICqHrcf//NWA1pXcc2IsbN4nygwwRVyvKxkY6Rtw5teCCkl0sCsi82+TCfbCZ9FRMJJ8JnKgOpbUW7QfvDDnuL4YOPHGcGb6mG413PZdflFdGgfcneEyLhYI8SdlU="
-    - secure: "LvKUGLEy9Z+olC07TRHMknoNa8O/7uTiieq4Zfb8o95MGBnjl63bPhLXqlU9pIM51lRzXzxBBCMUHJrVohmHK3ZvAwlfNpU5qWmXr2PXXUi71NNnMTmSYTG2xEOVO7jR3Er8LT7xSrX7SclmA6VbDPMb7UK5VxsxW2SGKGcSTmMy9d+ciIFFa1wbMFWOv3bY0MOCqliEeopbMM5tRe2iE+QFilPZ6QXXV3faRnxxk2DvB81IdMcGu2viHacCTByC95fNENKA2rNR7nGNMlf4Vn9twkVR5xUr6qCmfUOa0IMz8gHw25pVbD5HDDaRVf+gkCwVPjBV+ABstHtQednsPo5W1e5WsxJIMEOdKbVKRUzfxK8aC2U6XJrpnaO+TMoyjJVoyWvSBIaZbT966QGLV6MU264wxOoZbC7JsA3TPBIOVk0NMhprLdoEP5StAB5IzEG/YxhtuxEFmFVHQU5SJmx7wkFV3MjKxvw4vfSprdofxBev1ig9ZTfT0q6ssj7YAHWiGuvSEXSNAVOygogWlMpI0yWJPdxRGSriVVU0qPux6T5CpAH8g9yZhQwO8DNZSYzfxVV5LH6yPOGJo1riBS9sGBLz2EmsARk2b4SxTUeeXgOdxz3bxII/3HaJSiAILovNkwsCjwiPwMOFdTGoci2X0gPzvY+AVbcnrfDmp7k="
-
-branches:
-  except:
-    - gh-pages
-
-notifications:
-  email: false
-
-sudo: false
