--- assisted-inject-annotations-dagger2/build.gradle
@@ -1,10 +1,0 @@
-apply plugin: 'java-library'
-
-sourceCompatibility = JavaVersion.VERSION_1_6
-targetCompatibility = JavaVersion.VERSION_1_6
-
-dependencies {
-  api project(':assisted-inject-annotations')
-}
-
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
