--- assisted-inject-annotations-dagger2/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=Assisted Inject Dagger 2 Annotations
-POM_ARTIFACT_ID=assisted-inject-annotations-dagger2
-POM_PACKAGING=jar
