--- assisted-inject-annotations/build.gradle
@@ -1,6 +1,0 @@
-apply plugin: 'java-library'
-
-sourceCompatibility = JavaVersion.VERSION_1_6
-targetCompatibility = JavaVersion.VERSION_1_6
-
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
