--- assisted-inject-annotations/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=Assisted Inject Annotations
-POM_ARTIFACT_ID=assisted-inject-annotations
-POM_PACKAGING=jar
