--- assisted-inject-processor-dagger2/build.gradle
@@ -1,32 +1,0 @@
-import org.gradle.internal.jvm.Jvm
-
-apply plugin: 'java-library'
-apply plugin: 'org.jetbrains.kotlin.jvm'
-apply plugin: 'org.jetbrains.kotlin.kapt'
-
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-dependencies {
-  implementation project(':assisted-inject-annotations')
-  implementation project(':assisted-inject-annotations-dagger2')
-  implementation project(':assisted-inject-processor')
-  implementation deps.kotlin
-  implementation deps.javaPoet
-
-  compileOnly deps.autoService
-  kapt deps.autoService
-
-  compileOnly deps.incap.runtime
-  kapt deps.incap.processor
-
-  testImplementation deps.junit
-  testImplementation deps.truth
-  testImplementation deps.compileTesting
-  testImplementation deps.dagger
-  if (!Jvm.current().javaVersion.isJava9Compatible()) {
-    testImplementation files(Jvm.current().getToolsJar())
-  }
-}
-
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
