--- assisted-inject-processor-dagger2/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=Assisted Inject Dagger 2 Annotation Processor
-POM_ARTIFACT_ID=assisted-inject-processor-dagger2
-POM_PACKAGING=jar
