--- assisted-inject-processor/build.gradle
@@ -1,33 +1,0 @@
-import org.gradle.internal.jvm.Jvm
-
-apply plugin: 'java-library'
-apply plugin: 'org.jetbrains.kotlin.jvm'
-apply plugin: 'org.jetbrains.kotlin.kapt'
-
-sourceCompatibility = JavaVersion.VERSION_1_8
-targetCompatibility = JavaVersion.VERSION_1_8
-
-dependencies {
-  implementation project(':assisted-inject-annotations')
-  implementation deps.kotlin
-  api deps.javaPoet
-  implementation deps.autoCommon
-  implementation deps.guava
-
-  compileOnly deps.autoService
-  kapt deps.autoService
-
-  compileOnly deps.incap.runtime
-  kapt deps.incap.processor
-
-  testImplementation deps.junit
-  testImplementation deps.truth
-  testImplementation deps.compileTesting
-  testImplementation deps.inject
-  testImplementation deps.dagger
-  if (!Jvm.current().javaVersion.isJava9Compatible()) {
-    testImplementation files(Jvm.current().getToolsJar())
-  }
-}
-
-apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
