--- assisted-inject-processor/gradle.properties
@@ -1,3 +1,0 @@
-POM_NAME=Assisted Inject Annotation Processor
-POM_ARTIFACT_ID=assisted-inject-processor
-POM_PACKAGING=jar
