--- assisted-inject-sample-android/build.gradle
@@ -1,50 +1,0 @@
-apply plugin: 'com.android.application'
-
-android {
-  compileSdkVersion 28
-
-  defaultConfig {
-    minSdkVersion 14
-    targetSdkVersion 28
-
-    applicationId 'com.example.inject.assisted'
-  }
-
-  compileOptions {
-    sourceCompatibility JavaVersion.VERSION_1_8
-    targetCompatibility JavaVersion.VERSION_1_8
-  }
-
-  lintOptions {
-    checkReleaseBuilds false
-  }
-
-  signingConfigs {
-    assisted {
-      storeFile rootProject.file('signing.keystore')
-      storePassword 'assisted'
-      keyAlias 'assisted'
-      keyPassword 'assisted'
-    }
-  }
-
-  buildTypes {
-    debug {
-      signingConfig signingConfigs.assisted
-    }
-    release {
-      signingConfig signingConfigs.assisted
-
-      minifyEnabled true
-      proguardFiles(file('shrinker.pro'))
-    }
-  }
-}
-
-dependencies {
-  implementation deps.dagger
-  annotationProcessor deps.daggerCompiler
-
-  implementation project(':assisted-inject-annotations-dagger2')
-  annotationProcessor project(':assisted-inject-processor-dagger2')
-}
