--- assisted-inject-sample-android/shrinker.pro
@@ -1,4 +1,0 @@
--allowaccessmodification
--dontobfuscate
-
--checkdiscard class com.example.Greeter$Factory { *; }
