--- assisted-inject-sample-android/src/main/java/com/example/Exclamation.java
@@ -1,7 +1,0 @@
-package com.example;
-
-import javax.inject.Qualifier;
-
-@Qualifier
-public @interface Exclamation {
-}
