--- assisted-inject-sample-android/src/main/res/layout/main_view.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:padding="16dp"
-    >
-  <EditText
-      android:id="@+id/input"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:hint="Enter name"
-      />
-  <Button
-      android:id="@+id/show"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginTop="8dp"
-      android:hint="Show"
-      />
-  <TextView
-      android:id="@+id/output"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:layout_marginTop="8dp"
-      />
-</LinearLayout>
