+++ build.gradle
@@ -1,6 +1,6 @@
 buildscript {
   ext.versions = [
-      'kotlin': '1.3.41',
+      'kotlin': '1.4.31',
       'dagger': '2.33',
       'incap' : '0.2',
   ]
