+++ gradle.properties
@@ -1,12 +1,12 @@
 GROUP=com.squareup.inject
 VERSION_NAME=0.9.0-SNAPSHOT
 
-POM_DESCRIPTION=Assisted injection for JSR 330
+POM_DESCRIPTION=Constructor-inject views during layout inflation
 
-POM_URL=https://github.com/square/AssistedInject/
-POM_SCM_URL=https://github.com/square/AssistedInject/
-POM_SCM_CONNECTION=scm:git:git://github.com/square/AssistedInject.git
-POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/square/AssistedInject.git
+POM_URL=https://github.com/square/InflationInjection/
+POM_SCM_URL=https://github.com/square/InflationInjection/
+POM_SCM_CONNECTION=scm:git:git://github.com/square/InflationInjection.git
+POM_SCM_DEV_CONNECTION=scm:git:ssh://git@github.com/square/InflationInjection.git
 
 POM_LICENCE_NAME=The Apache Software License, Version 2.0
 POM_LICENCE_URL=http://www.apache.org/licenses/LICENSE-2.0.txt
